Lift the Stats endpoints out of web_server.py
21 routes to api/stats.py (-433 lines). Worker getters because both stats
workers are legitimately None when their boot failed. Url-map verified.
